cleanup and deps updated
